Uploaded First Version of JAPC
This version is very limited and doesn't have as many features as version 2, but it gets the job done. This doesn't make a .txt file and you have to continously start the program again if you want to make another character. It also doesn't have fully random names, so it generates a full name rather than a random first and last name.